+++ pom.xml
@@ -77,7 +77,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-autoconfigure</artifactId>
-            <version>3.0.11-SNAPSHOT</version>
         </dependency>
     </dependencies>
