+++ src/main/java/com/HocusPocus/Simulation/simulationDb/repo/SimulationRepo.java
@@ -1,4 +1,4 @@
-package com.HocusPocus.Simulation.repo;
+package com.HocusPocus.Simulation.simulationDb.repo;
 
 import com.HocusPocus.Simulation.simulationDb.SimulationResults;
 import org.springframework.data.jpa.repository.JpaRepository;
