+++ src/test/java/com/HocusPocus/Simulation/SimulationTests.java
@@ -1,7 +1,7 @@
 package com.HocusPocus.Simulation;
 
-import com.HocusPocus.Simulation.repo.SimulationRepo;
 import com.HocusPocus.Simulation.simulationDb.SimulationResults;
+import com.HocusPocus.Simulation.simulationDb.repo.SimulationRepo;
 import io.restassured.RestAssured;
 import org.junit.Before;
 import org.junit.Test;
